Added redux-store, changed the app structure
