+++ src/App.js
@@ -1,41 +1,28 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
+import React from "react";
 import { Switch, Route } from "react-router-dom";
 
 import Header from "./components/Header";
 import Restaurant from "./components/Restaurant";
-import FilterComponent from "./components/List/FilterComponent";
+import Restaurants from "./components/Restaurants";
 
 import "./index.scss";
 
-const apiLink = "https://redi-final-restaurants.herokuapp.com/restaurants";
-
-function App() {
-  const [content, setContent] = useState([]);
-  useEffect(() => {
-    axios
-      .get(apiLink)
-      .then((res) => {
-        setContent(res.data.results);
-      })
-      .catch((error) => console.log(error));
-  }, []);
-
+const App = () => {
   return (
     <div className="wrapper">
       <Header />
       <main>
         <Switch>
           <Route path="/restaurants/:name">
-            <Restaurant rests={content} />
+            <Restaurant />
           </Route>
           <Route path="/">
-            <FilterComponent rests={content} />
+            <Restaurants />
           </Route>
         </Switch>
       </main>
     </div>
   );
-}
+};
 
 export default App;
