+++ src/redux-store/actionTypes.js
@@ -1,0 +1,6 @@
+export const actionTypes = {
+  startLoading: "START_LOADING",
+  loadSuccess: "LOAD_SUCCESS",
+  loadError: "LOAD_ERROR",
+  filterAction: "FILTER",
+};
