+++ src/redux-store/store.js
@@ -1,0 +1,6 @@
+import { createStore } from "redux";
+import restaurantsReducer from "./reducer";
+
+const store = createStore(restaurantsReducer);
+
+export default store;
